v0.0.1 All system working together after first refactoring WIP: There are some bugs with Octree LODs and creating the mesh from a system isn't working

+++ Assets/Modules/Terrain Generator Module/Scripts/Models/TerrainGeneratorData.cs
@@ -14,6 +14,6 @@ namespace TerrainGenerator
         [Range(minSize, maxSize)]
         public float size;
 
-        public int[] depthResolution;
+        public ushort[] depthResolution;
     }
 }
